+++ changelog.md
@@ -17,7 +17,8 @@
 - Performance improvements
 
 # Version: 0.1.6
-- Changes in which properties belong to style and which to type to make it more usable
-  Sorry for the breaking changes
+- Changes in which properties belong to style and which to type to make it more usable 
+    - Grades are now listed under style in stead of type
+- When editing a style or type the active type, style is now used to generate the preview 
 - Faster loading speeds
 - UI improvements